Moved locales out of translations folder to config
