+++ components/Jobs/JobPostBody.tsx
@@ -8,7 +8,7 @@ import {
 import Button from '../Buttons/Button';
 import { useLocale, useTranslations } from 'next-intl';
 import Link from '../Link/Link';
-import { localeToDateLocale } from '@/translations/locales';
+import { localeToDateLocale } from '@/config/locales';
 
 const PostBody = ({ jobPost }: { jobPost?: JobPostingWithId }) => {
   // Creates the body of the post constaining the title, body text and also handles editing logic
